+++ angel_orm_postgres/CHANGELOG.md
@@ -1,5 +1,6 @@
 # 1.1.0-beta.1
 * Improvements in how transactions are handled; rethrow failed exceptions after rolling back.
+* Remove deprecated executor classes.
 
 # 1.1.0-beta
 * Updates for `package:angel_orm@2.1.0-beta`.